Move mimir MapFuncPlanner to its own file.
