+++ mimir/src/main/scala/quasar/mimir/Mimir.scala
@@ -31,11 +31,9 @@ import quasar.fs.mount._
 import quasar.qscript._
 
 import quasar.blueeyes.json.{JNum, JValue}
-import quasar.precog.common.{Path, RValue}
+import quasar.precog.common.Path
 import quasar.yggdrasil.vfs.ResourceError
 import quasar.yggdrasil.bytecode.JType
-import quasar.yggdrasil.table.cf.util.Undefined
-import quasar.yggdrasil.table.cf.math
 
 import fs2.Stream
 import fs2.async.mutable.{Queue, Signal}
@@ -158,6 +156,8 @@ object Mimir extends BackendModule with Logging {
 //def interpretM[M[_], F[_], A, B](f: A => M[B], φ: AlgebraM[M, F, B]): AlgebraM[M, CoEnv[A, F, ?], B]
 // f.cataM(interpretM)
 
+    val mapFuncPlanner = new MapFuncPlanner[T, M, Backend]
+
     lazy val planQST: AlgebraM[Backend, QScriptTotal[T, ?], Repr] =
       _.run.fold(
         planQScriptCore,
@@ -177,130 +177,13 @@ object Mimir extends BackendModule with Logging {
                       _ => ???,   // Read[AFile]
                       _ => ???))))))))    // DeadEnd
 
-    def planMapFunc(cake: Cake): AlgebraM[Backend, MapFuncCore[T, ?], cake.trans.TransSpec1] = {
-      case MapFuncsCore.Undefined() =>
-        (cake.trans.Map1[cake.trans.Source1](cake.trans.TransSpec1.Id, Undefined): cake.trans.TransSpec1).point[M].liftB
-
-      case MapFuncsCore.Constant(ejson) =>
-        // EJson => Data => JValue => RValue => Table
-        val data: Data = ejson.cata(Data.fromEJson)
-        val jvalue: JValue = JValue.fromData(data)
-        val rvalue: RValue = RValue.fromJValue(jvalue)
-        cake.trans.transRValue(rvalue, cake.trans.TransSpec1.Id).point[Backend]
-
-      case MapFuncsCore.JoinSideName(_) => ??? // should never be received
-
-      case MapFuncsCore.Length(a1) => ???
-
-      case MapFuncsCore.ExtractCentury(a1) => ???
-      case MapFuncsCore.ExtractDayOfMonth(a1) => ???
-      case MapFuncsCore.ExtractDecade(a1) => ???
-      case MapFuncsCore.ExtractDayOfWeek(a1) => ???
-      case MapFuncsCore.ExtractDayOfYear(a1) => ???
-      case MapFuncsCore.ExtractEpoch(a1) => ???
-      case MapFuncsCore.ExtractHour(a1) => ???
-      case MapFuncsCore.ExtractIsoDayOfWeek(a1) => ???
-      case MapFuncsCore.ExtractIsoYear(a1) => ???
-      case MapFuncsCore.ExtractMicroseconds(a1) => ???
-      case MapFuncsCore.ExtractMillennium(a1) => ???
-      case MapFuncsCore.ExtractMilliseconds(a1) => ???
-      case MapFuncsCore.ExtractMinute(a1) => ???
-      case MapFuncsCore.ExtractMonth(a1) => ???
-      case MapFuncsCore.ExtractQuarter(a1) => ???
-      case MapFuncsCore.ExtractSecond(a1) => ???
-      case MapFuncsCore.ExtractTimezone(a1) => ???
-      case MapFuncsCore.ExtractTimezoneHour(a1) => ???
-      case MapFuncsCore.ExtractTimezoneMinute(a1) => ???
-      case MapFuncsCore.ExtractWeek(a1) => ???
-      case MapFuncsCore.ExtractYear(a1) => ???
-      case MapFuncsCore.Date(a1) => ???
-      case MapFuncsCore.Time(a1) => ???
-      case MapFuncsCore.Timestamp(a1) => ???
-      case MapFuncsCore.Interval(a1) => ???
-      case MapFuncsCore.StartOfDay(a1) => ???
-      case MapFuncsCore.TemporalTrunc(part, a1) => ???
-      case MapFuncsCore.TimeOfDay(a1) => ???
-      case MapFuncsCore.ToTimestamp(a1) => ???
-      case MapFuncsCore.Now() => ???
-
-      case MapFuncsCore.TypeOf(a1) => ???
-
-      case MapFuncsCore.Negate(a1) =>
-        (cake.trans.Map1[cake.trans.Source1](a1, math.Negate): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Add(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.Add.f2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Multiply(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.Mul.f2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Subtract(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.Sub.f2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Divide(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.Div.f2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Modulo(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.Mod.f2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Power(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.Pow.f2): cake.trans.TransSpec1).point[M].liftB
-
-      case MapFuncsCore.Not(a1) => ???
-      case MapFuncsCore.Eq(a1, a2) =>
-        (cake.trans.Equal[cake.trans.Source1](a1, a2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Neq(a1, a2) => ???
-      case MapFuncsCore.Lt(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.Lt.f2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Lte(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.LtEq.f2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Gt(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.Gt.f2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Gte(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.GtEq.f2): cake.trans.TransSpec1).point[M].liftB
-
-      case MapFuncsCore.IfUndefined(a1, a2) => ???
-      case MapFuncsCore.And(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.And.f2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Or(a1, a2) =>
-        (cake.trans.Map2[cake.trans.Source1](a1, a2, cake.Library.Infix.Or.f2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.Between(a1, a2, a3) => ???
-      case MapFuncsCore.Cond(a1, a2, a3) => ???
-
-      case MapFuncsCore.Within(a1, a2) => ???
-
-      case MapFuncsCore.Lower(a1) => ???
-      case MapFuncsCore.Upper(a1) => ???
-      case MapFuncsCore.Bool(a1) => ???
-      case MapFuncsCore.Integer(a1) => ???
-      case MapFuncsCore.Decimal(a1) => ???
-      case MapFuncsCore.Null(a1) => ???
-      case MapFuncsCore.ToString(a1) => ???
-      case MapFuncsCore.Search(a1, a2, a3) => ???
-      case MapFuncsCore.Substring(string, from, count) => ???
-
-      // FIXME detect constant cases so we don't have to always use the dynamic variants
-      case MapFuncsCore.MakeArray(a1) =>
-        (cake.trans.WrapArray[cake.trans.Source1](a1): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.MakeMap(key, value) =>
-        (cake.trans.WrapObjectDynamic[cake.trans.Source1](key, value): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.ConcatArrays(a1, a2) =>
-        (cake.trans.OuterArrayConcat[cake.trans.Source1](a1, a2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.ConcatMaps(a1, a2) =>
-        (cake.trans.OuterObjectConcat[cake.trans.Source1](a1, a2): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.ProjectIndex(src, index) =>
-        (cake.trans.DerefArrayDynamic[cake.trans.Source1](src, index): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.ProjectField(src, field) =>
-        (cake.trans.DerefObjectDynamic[cake.trans.Source1](src, field): cake.trans.TransSpec1).point[M].liftB
-      case MapFuncsCore.DeleteField(src, field) => ???
-
-      case MapFuncsCore.Meta(a1) => ???
-
-      case MapFuncsCore.Range(from, to) => ???
-
-      // FIXME if fallback is not undefined don't throw this away
-      case MapFuncsCore.Guard(_, _, a2, _) => a2.point[Backend]
-    }
-
     lazy val planQScriptCore: AlgebraM[Backend, QScriptCore[T, ?], Repr] = {
       case qscript.Map(src, f) =>
         for {
           trans <- f.cataM[Backend, src.P.trans.TransSpec1](
-            interpretM(κ(src.P.trans.TransSpec1.Id.point[Backend]), planMapFunc(src.P)))
+            interpretM(
+              κ(src.P.trans.TransSpec1.Id.point[Backend]),
+              mapFuncPlanner.plan(src.P)))
         } yield Repr(src.P)(src.table.transform(trans))
 
       case qscript.LeftShift(src, struct, id, repair) => ???
@@ -310,7 +193,9 @@ object Mimir extends BackendModule with Logging {
       case qscript.Filter(src, f) =>
         for {
           trans <- f.cataM[Backend, src.P.trans.TransSpec1](
-            interpretM(κ(src.P.trans.TransSpec1.Id.point[Backend]), planMapFunc(src.P)))
+            interpretM(
+              κ(src.P.trans.TransSpec1.Id.point[Backend]),
+              mapFuncPlanner.plan(src.P)))
         } yield Repr(src.P)(src.table.transform(src.P.trans.Filter(src.P.trans.TransSpec1.Id, trans)))
 
       case qscript.Union(src, lBranch, rBranch) => ???
